langchain/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dataclasses-json==0.6.1
exceptiongroup==1.1.3
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.12.0
google-api-python-client==2.100.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-cloud-aiplatform==1.34.0
google-cloud-bigquery==3.12.0
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
install==1.3.5
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
langchain==0.0.307
langsmith==0.0.42
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.24.4
packaging==23.2
pandas==2.0.3
Pillow==10.0.1
pkgutil-resolve-name==1.3.10
proto-plus==1.22.3
protobuf==4.24.4
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic-core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.4
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.21
streamlit==1.27.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing-extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.1
uritemplate==4.1.1
urllib3==2.0.6
validators==0.22.0
yarl==1.9.2
youtube-search==2.1.2
zipp==3.17.0
